pyproject.toml
[project]
name = "ru-gost-crypto-lab"
version = "0.1.0"
description = "Laboratory stand for testing GOST cryptography infrastructure"
authors = [
    {name = "Ranas Mukminov", email = "info@run-as-daemon.ru"},
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "bandit>=1.7",
    "types-PyYAML",
    "types-requests",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
